Improve ed_prev_line/ed_next_line, add ed_smart_*_line
Assume we were in a horizontal position +20 and moved into
a shorter line. Track the initial position so that once
moved into a longer line, we are once again in +20
(provided that the sequence consists of vertical movements only).

Ed_smart_prev_line moves unless already on the first line;
in that case it performs ed_prev_history action. Similar
story with ed_smart_next_line.